+++ src/ide/MegaSCSI.hh
@@ -26,7 +26,7 @@ public:
 	void serialize(Archive& ar, unsigned version);
 
 private:
-	[[nodiscard]]unsigned getSramSize() const;
+	[[nodiscard]] size_t getSramSize() const;
 	void setSRAM(unsigned region, byte block);
 
 private:
